refactor: rename configGuards.ts to providerQuirks.ts
Better describes the purpose - handling provider-specific parameter
requirements and quirks.
