+++ packages/agent-sdk-ts/src/sdk/llm/providerQuirks.ts
